Rename keyboards with uppercase letters
