+++ data/schemas/definitions.jsonschema
@@ -68,20 +68,8 @@
         "type": "number"
     },
     "keyboard": {
-        "oneOf": [
-            {
-                "type": "string",
-                "enum": [
-                    "converter/numeric_keypad_IIe",
-                    "emptystring/NQG",
-                    "maple_computing/christmas_tree/V2017"
-                ]
-            },
-            {
-                "type": "string",
-                "pattern": "^[0-9a-z][0-9a-z_/]*$"
-            }
-        ]
+        "type": "string",
+        "pattern": "^[0-9a-z][0-9a-z_/]*$"
     },
     "mcu_pin_array": {
         "type": "array",
