+++ keyboards/converter/numeric_keypad_iie/numeric_keypad_iie.c
@@ -13,4 +13,4 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "numeric_keypad_IIe.h"
+#include "numeric_keypad_iie.h"
