+++ keyboards/converter/numeric_keypad_iie/readme.md
@@ -11,6 +11,6 @@ Hardware Supported: Internal Numeric Keypad IIe PCB (Apple part 820.0081.c), ATm
 
 Make example for this keyboard (after setting up your build environment):
 
-    make converter/numeric_keypad_IIe:default
+    make converter/numeric_keypad_iie:default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
