--- keyboards/emptystring/NQG/NQG.c
@@ -1,1 +1,0 @@
-#include "NQG.h"
