+++ keyboards/emptystring/nqg/config.h
