+++ keyboards/emptystring/nqg/keymaps/default/keymap.c
