+++ keyboards/emptystring/nqg/keymaps/default/rules.mk
