+++ keyboards/emptystring/nqg/nqg.c
@@ -1,0 +1,1 @@
+#include "nqg.h"
