--- keyboards/maple_computing/christmas_tree/V2017/V2017.c
@@ -1,1 +1,0 @@
-#include "V2017.h"
