+++ keyboards/maple_computing/christmas_tree/readme.md
@@ -9,6 +9,6 @@ Keyboard Maintainer: [That-Canadian](https://github.com/That-Canadian)
 
 Make example for this keyboard (after setting up your build environment):
 
-    make maple_computing/christmas_tree/V2017:default
+    make maple_computing/christmas_tree/v2017:default
 
 See [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) then the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information.
