+++ keyboards/maple_computing/christmas_tree/rules.mk
@@ -17,4 +17,4 @@ BACKLIGHT_ENABLE = yes      # Enable keyboard backlight functionality
 AUDIO_ENABLE = no           # Audio output
 RGBLIGHT_ENABLE = no        # Enable WS2812 RGB underlight.
 
-DEFAULT_FOLDER = maple_computing/christmas_tree/V2017
+DEFAULT_FOLDER = maple_computing/christmas_tree/v2017
